requirements.txt
anyio==4.6.0
autopep8==2.3.1
cachetools==5.5.1
certifi==2024.8.30
charset-normalizer==3.4.1
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
numpy==2.2.2
packaging==24.1
pandas==2.2.3
pluggy==1.5.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pyparsing==3.2.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
telegram==0.0.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.3.0
